+++ http/handler.go
@@ -6,11 +6,10 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/hashicorp/errwrap"
+	"github.com/hashicorp/vault/helper/duration"
 	"github.com/hashicorp/vault/helper/jsonutil"
 	"github.com/hashicorp/vault/logical"
 	"github.com/hashicorp/vault/vault"
@@ -167,23 +166,14 @@ func requestWrapTTL(r *http.Request, req *logical.Request) (*logical.Request, er
 	}
 
 	// If it has an allowed suffix parse as a duration string
-	if strings.HasSuffix(wrapTTL, "s") || strings.HasSuffix(wrapTTL, "m") || strings.HasSuffix(wrapTTL, "h") {
-		dur, err := time.ParseDuration(wrapTTL)
-		if err != nil {
-			return req, err
-		}
-		req.WrapTTL = dur
-	} else {
-		// Parse as a straight number of seconds
-		seconds, err := strconv.ParseInt(wrapTTL, 10, 64)
-		if err != nil {
-			return req, err
-		}
-		req.WrapTTL = time.Duration(seconds) * time.Second
+	dur, err := duration.ParseDurationSecond(wrapTTL)
+	if err != nil {
+		return req, err
 	}
-	if int64(req.WrapTTL) < 0 {
+	if int64(dur) < 0 {
 		return req, fmt.Errorf("requested wrap ttl cannot be negative")
 	}
+	req.WrapTTL = dur
 
 	return req, nil
 }