+++ logical/framework/field_data.go
@@ -3,10 +3,8 @@ package framework
 import (
 	"encoding/json"
 	"fmt"
-	"strconv"
-	"strings"
-	"time"
 
+	"github.com/hashicorp/vault/helper/duration"
 	"github.com/mitchellh/mapstructure"
 )
 
@@ -163,21 +161,11 @@ func (d *FieldData) getPrimitive(
 		case float64:
 			result = int(inp)
 		case string:
-			// Look for a suffix otherwise its a plain second value
-			if strings.HasSuffix(inp, "s") || strings.HasSuffix(inp, "m") || strings.HasSuffix(inp, "h") {
-				dur, err := time.ParseDuration(inp)
-				if err != nil {
-					return nil, true, err
-				}
-				result = int(dur.Seconds())
-			} else {
-				// Plain integer
-				val, err := strconv.ParseInt(inp, 10, 64)
-				if err != nil {
-					return nil, true, err
-				}
-				result = int(val)
+			dur, err := duration.ParseDurationSecond(inp)
+			if err != nil {
+				return nil, true, err
 			}
+			result = int(dur.Seconds())
 		case json.Number:
 			valInt64, err := inp.Int64()
 			if err != nil {